Restructure code for new release
This restructures the code:
- Rename addon.py into ishelper.py
- Move api.py and config.py out of the global module namespace
- Fix tests to accomodate this
