+++ Makefile
@@ -9,7 +9,7 @@ git_branch = $(shell git rev-parse --abbrev-ref HEAD)
 git_hash = $(shell git rev-parse --short HEAD)
 
 zip_name = $(name)-$(version)-$(git_branch)-$(git_hash).zip
-include_files = addon.py addon.xml changelog.txt LICENSE.txt README.md lib/ resources/
+include_files = addon.xml changelog.txt default.py LICENSE.txt README.md lib/ resources/
 include_paths = $(patsubst %,$(name)/%,$(include_files))
 exclude_files = \*.new \*.orig \*.pyc \*.pyo
 zip_dir = $(name)/
@@ -43,6 +43,7 @@ addon: clean
 
 unit: clean
 	@echo -e "$(white)=$(blue) Starting unit tests$(reset)"
+	python default.py
 	python -m unittest discover
 #	coverage run -m unittest discover
