+++ lib/inputstreamhelper/api.py
