+++ lib/inputstreamhelper/config.py
