+++ tox.ini
@@ -16,6 +16,7 @@ max-line-length = 160
 ignore = E129,FI13,FI50,FI51,FI53,FI54,W503
 require-code = True
 min-version = 2.7
+exclude = .tox,experiment
 
 [pytest]
 filterwarnings = default
